requirements.txt
tensorflow-rl 
keras-rl 
gymnasium[atari] 
Pillow